requirements.txt
# Python master - Requirements
# Учебная платформа для изучения Python

# FastAPI и веб
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9

# База данных
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Валидация
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Безопасность
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Для работы с датами
python-dateutil==2.8.2

# Для сериализации
orjson==3.9.13

# Работа с кодом (подсветка синтаксиса в редакторе)
pygments==2.17.2

# Aiogram для Telegram уведомлений (опционально)
# aiogram==3.4.1